requirements.txt
pywin32
PyYAML
rich
InquirerPy
playwright
